🐛 fix: split the configuration file to be used only on the server-side when using the Firebase SDK.
